requirements-cpu.txt
numpy
scipy
laspy
Pillow
cython
shapely
matplotlib
scikit-image
tensorflow==1.14.0
keras==2.0.8
opencv-python
h5py
imgaug
IPython[all]
